requirements_prod.txt
Flask-Collect >=1.3.2, <2
waitress >= 1.1.0, <2
